--- bot/__init__.py
@@ -1,7 +1,0 @@
-from bot.__main__ import send_ad_to_telegram, edit_ad_in_telegram, send_comment_for_ad_to_telegram
-
-__all__ = [
-    'send_ad_to_telegram',
-    'edit_ad_in_telegram',
-    'send_comment_for_ad_to_telegram',
-]
