+++ docker-compose.yaml
@@ -41,4 +41,4 @@ services:
       - TZ=Europe/Istanbul
     volumes:
       - ./:/opt/app
-    command: ['poetry', 'run', 'python', '-m', 'bot']
+    command: ['poetry', 'run', 'python', '-m', 'telegram']
