+++ telegram/__main__.py
@@ -1,0 +1,4 @@
+from telegram.bot import bot
+
+if __name__ == '__main__':
+    bot.infinity_polling()
