+++ telegram/models.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class TelegramIdAd(BaseModel):
+    _id: str
+    telegram_channel_message_id: str
+    telegram_chat_message_id: str
